requirements.txt
python-telegram-bot
google-generativeai
flask
python-dotenv
redis
